+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [2.5.3]
+### Added
+- Kyverno policy test script
+- Kyverno cli to bb-ci image
+
+### Updated
+- Helm version to `3.10.2` in bb-ci image
+- Removed `conftest` from bb-ci image
+
 ## [2.5.2]
 ### Added
 - Added BB CI registry to allowed list of registries